test(landing): correct test_runtime_bridge_identity markers (git_repo present, fast removed)
Pre-existing main breakage folded per the maintainer landing runbook
(docs/development/pr-landing.md step 4/5) while landing PR #3175; PR #3175
does not touch this file. Reproduced identically on merge-base abca7ec96.

CI's arch-adversarial (arch_shard_3) was red on two architectural gates in
tests/architectural/test_pytest_marker_correctness.py:
  - test_subprocess_git_users_must_carry_git_repo_marker (Rule 1)
  - test_fast_marker_must_not_apply_to_subprocess_users (Rule 2)

Both named tests/runtime/test_runtime_bridge_identity.py. Confirmed the file
was NOT a uniform subprocess/git user: 6 of 8 tests (TestResolveMissionUlid,
TestWrapWithDecisionGitLogIdentityContract) only use unittest.mock.patch, no
subprocess calls anywhere; only TestMissionRoutesThroughCoordinationCorruptMeta
(2 tests, the WP08/#3155 corrupt-meta regression) drives real git via
subprocess.run(["git", ...]) through its _init_repo/_seed helpers. A wholesale
marker swap (drop fast, add git_repo on the whole file) would have silently
removed 6 genuinely fast, mock-only tests from the inner dev loop's -m fast
profile.

Per Rule 2's fix-hint split option, moved the corrupt-meta class into a new
sibling file, tests/runtime/test_runtime_bridge_identity_git_repo.py, marked
[unit, git_repo] (no fast). The original file keeps [unit, fast] with only
the mock-only tests remaining.
